requirements.txt
aniso8601==7.0.0
backcall==0.2.0
bcrypt==3.2.0
black==21.9b0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
debugpy==1.4.3
decorator==5.1.0
entrypoints==0.3
Flask==2.0.1
Flask-GraphQL==2.0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
greenlet==1.1.2
idna==3.2
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jupyter-client==7.0.5
jupyter-core==4.8.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mongoengine==0.23.1
mypy-extensions==0.4.3
nest-asyncio==1.5.1
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.0
promise==2.3
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.10.0
pymongo==3.12.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pyzmq==22.3.0
regex==2021.9.24
requests==2.26.0
Rx==1.6.1
six==1.16.0
spotipy==2.19.0
SQLAlchemy==1.4.25
tomli==1.2.1
tornado==6.1
traitlets==5.1.0
typing-extensions==3.10.0.2
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.1
youtube-dl==2021.6.6
